requirements.txt
flask
flask_sqlalchemy
flask_login
flask_socketio
pywebview
pythonping
passlib
bcrypt
requests
python-dateutil
pytz